requirements.txt
plotly==5.24.1
yfinance==0.2.40
google-generativeai
openai
phidata
googlesearch-python
pycountry
streamlit
